requirements.txt
beautifulsoup4==4.14.3
lyricsgenius==3.7.6
mutagen==1.47.0
python-dotenv==1.2.1
requests==2.32.5
spotipy==2.25.2
